light-saber: resized handle lid and blades.
Reducing the blade top/bottom difference made the
whole sword smaller rather than bigger; making it
now a 1.6 instead of the baseline 1.4 to see if it
results in a longer expansion.

Also, the thread calculation was wrong and it wasn't
fitting the handle lid in the handle thread. Adjusted
that part as well.