feat: added organizer models, reworked .stl export
Organizer models were implemented to be used as an organizer box
with several shapes and sizes.

Reworked how models are generated, taking advantage of the Openscad
parameter presets in order to generated several STL variations of
the same SCAD file.

Also reworked how PNG preview images are generated to align with the
preset variations as well.